+++ tests/_bundles/security/serializers/user_serializer.py
@@ -9,6 +9,10 @@ NON_ALPHANUMERIC_RE = re.compile(r'[^\w]')
 
 
 class UserSerializer(BaseUserSerializer):
+    username = ma.String(required=True)
+    first_name = ma.String(required=True)
+    last_name = ma.String(required=True)
+
     @ma.validates('username')
     def validate_username(self, username):
         if re.search(NON_ALPHANUMERIC_RE, username):